+++ PortfolioManagerWASM/Models/DTOs/PortfolioAssetDto.cs
@@ -6,6 +6,6 @@ namespace PortfolioManagerWASM.Models.DTOs
     public class PortfolioAssetDto
     {
         public FinancialAssetDto FinancialAssetDto { get; set; }
-        public float AllocationPercentage { get; set; }
+        public int AllocationPercentage { get; set; }
     }
 }
